test: extract shared WebAuthn E2E helpers and harden Keycloak startup
Addresses the two low-severity review findings on PR #93 (#90).

- Extract setupVirtualAuthenticator, addVirtualAuthenticator, and
  getCredentialIds into playwright/src/utils/webauthn.ts and export
  them from utils/index.ts. The step-up-oidc, step-up-flow, and
  mfa-flow specs now import these instead of each carrying a local
  copy (mfa-flow previously inlined the authenticator setup).

- startKeycloak now removes the container it started if waitForMetadata
  times out. Playwright's separate globalTeardown is not guaranteed to
  run when globalSetup throws, so a readiness timeout could otherwise
  leave the container and ownership marker behind, which the next run's
  "already running but not answering" guard would reject until a manual
  docker rm -f.
